Creating Intro and Board components
