+++ src/domains/game/components/Board/styles.ts
@@ -1,0 +1,27 @@
+import styled from 'styled-components/native';
+
+export const MainView = styled.View`
+  flex: 1;
+`;
+
+export const HeaderView = styled.View`
+  flex: 2;
+`;
+
+export const SequencerView = styled.View`
+  flex: 5;
+  background-color: #000000;
+  padding: 5px;
+`;
+
+export const TimerView = styled.View`
+  flex: 1.1;
+  flex-direction: row;
+  background-color: #000000;
+`;
+
+export const PlayerView = styled.View`
+  background-color: #000000;
+  flex: 5;
+  padding: 5px 5px 20px;
+`;
