+++ src/domains/game/components/Intro/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+export const Intro = () => {
+  console.log('intro...');
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Intro</Text>
+    </View>
+  );
+};
