+++ src/screens/Game/index.tsx
@@ -1,14 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import type { StackNavigationProp } from '@react-navigation/stack';
-import { MainView, HeaderView, PlayerView, SequencerView, TimerView } from 'screens/Game/styles';
-import { useGame } from 'domains/game/data/hooks/useGame';
-import { Sequencer } from 'domains/game/components/Sequencer';
-import { KeyPad } from 'domains/player/components/KeyPad';
-import { StatusBar } from 'domains/game/components/StatusBar';
-import { Header } from 'domains/game/components/Header';
+import { Board } from 'domains/game/components/Board';
+import { Intro } from 'domains/game/components/Intro';
 import type { RootStackParamList } from 'screens/types';
-import { logSequenceOutput } from 'helpers/logSequenceOutput';
-import { GameOverModal } from 'domains/shell/components/GameOverModal';
 
 type GameScreenNavigationProp = StackNavigationProp<RootStackParamList, 'Home'>;
 
@@ -17,62 +11,10 @@ type Props = {
 };
 
 export const Game = ({ navigation }: Props) => {
-  const {
-    addPlayerMove,
-    attemptsLeft,
-    currentSequence,
-    hasQuit,
-    isGameOver,
-    level,
-    playerSequence,
-    quit,
-    rightSequences,
-    score,
-    speed,
-    start,
-  } = useGame();
-  const [choseExit, setChoseExit] = useState(false);
-
-  logSequenceOutput('CURRENT', currentSequence);
-  logSequenceOutput('PLAYER ', playerSequence);
-  console.log(
-    `SPEED: ${speed}, SCORE: ${score}, LEVEL: ${level},  NAILED ${rightSequences}, ATTEMPTS ${attemptsLeft}`,
-  );
-  console.log('');
-
-  const goHome = () => {
-    setChoseExit(true);
+  const navigateHome = () => {
     navigation.navigate('Home');
   };
 
-  const tryAgain = () => {
-    start();
-  };
-
-  const exitGame = () => {
-    quit();
-    navigation.navigate('Home');
-  };
-
-  return (
-    <MainView>
-      <GameOverModal
-        isOpen={!hasQuit && isGameOver && !choseExit}
-        onExit={goHome}
-        onTryAgain={tryAgain}
-      />
-      <HeaderView>
-        <Header isGameOver={isGameOver} onExit={exitGame} />
-      </HeaderView>
-      <SequencerView>
-        <Sequencer />
-      </SequencerView>
-      <TimerView>
-        <StatusBar />
-      </TimerView>
-      <PlayerView>
-        <KeyPad onKeyPress={addPlayerMove} />
-      </PlayerView>
-    </MainView>
-  );
+  // return <Intro />;
+  return <Board onGoHome={navigateHome} />;
 };
